requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
mysql-connector-python>=8.3.0
pandas>=2.0.0
